Gose

+++ src/components/Feed/MyWriteFeed.vue
@@ -18,7 +18,7 @@
               a.columns.is-mobile.group-small-list-group(@click="goGroup(group.pk)")
                 article.media.group-small-list
                   figure.media-left
-                    p.image.is-32x32
+                    p.image.is-32x32.group-img-small-wrapper
                       img.group-img-small(:src='group.profile_img')
                 p.group-small-name {{group.name}}
             .card-content
@@ -31,44 +31,45 @@
 
           //- feed 영역
           .column.is-9
-            //- div.feed-box(v-show="post_data.length <= 0")
-              .card
-                .card-content
-                  .content
-                    | 그룹에 재미있는 이야기를 써보세요.
+            .feed-box-wrapper
+              div.feed-box(v-show="data_list.length <= 0")
+                .card
+                  .card-content
+                    .content
+                      | 그룹에 재미있는 이야기를 써보세요.
 
 
-            //- 컨텐츠가 들어간 글
-            div.feed-box
-              .card(v-for = "data in data_list")
-                header.card-header
-                  a.card-header-title.group-name(@click.prevent ="goGroup(data.group.pk)") 
-                    | &nbsp;  
-                    | &nbsp;  
-                    span.icon.icon-bond
-                      img(src="../../assets/btn-bond-normal.svg")
-                    | &nbsp;  
-                    | {{data.group.name}}
-                .card-content
-                  article.media
-                    .media-left
-                      figure.image.is-64x64.img-user
-                        img.user-img(:src='data.author.profile_img', alt='Image')
-                    .media-content
-                      p.title.is-4.user-name {{data.author.nickname}}
-                      p.subtitle.is-6 {{calcDate (data.created_date)}}
-                    button.delete(@click="deletePost(data.pk)")
-                  //- 글 (최상위)
-                  .content {{data.content}}
-                  //- 이미지 - 1개일 때
-                  .content
-                    figure.image
-                      img(:src='data.image')
-              .columns
-                .column
-                  nav.pagination.is-centered
-                    button.pagination-previous.pagination-btn(@click="prevPage()" :disabled='pagination.prev === null') 이전 페이지
-                    button.pagination-next.pagination-btn(@click="nextPage()" :disabled='pagination.next === null') 다음 페이지                        
+              //- 컨텐츠가 들어간 글
+              div.feed-box
+                .card(v-for = "data in data_list")
+                  header.card-header
+                    a.card-header-title.group-name(@click.prevent ="goGroup(data.group.pk)") 
+                      | &nbsp;  
+                      | &nbsp;  
+                      span.icon.icon-bond
+                        img(src="../../assets/btn-bond-normal.svg")
+                      | &nbsp;  
+                      | {{data.group.name}}
+                  .card-content
+                    article.media
+                      .media-left
+                        figure.image.is-64x64.img-user-64.is-1by1
+                          img.img-user-profile(:src='data.author.profile_img', alt='Image')
+                      .media-content
+                        p.title.is-4.user-name {{data.author.nickname}}
+                        p.subtitle.is-6 {{calcDate (data.created_date)}}
+                      button.delete(@click="deletePost(data.pk)")
+                    //- 글 (최상위)
+                    .content {{data.content}}
+                    //- 이미지 - 1개일 때
+                    .content
+                      figure.image
+                        img(:src='data.image')
+            .columns
+              .column
+                nav.pagination.is-centered
+                  button.pagination-previous.pagination-btn(@click="prevPage()" :disabled='pagination.prev === null') 이전 페이지
+                  button.pagination-next.pagination-btn(@click="nextPage()" :disabled='pagination.next === null') 다음 페이지                        
 
       main-footer
       MakingGroupModal(ref="my_modal" close_message="close lightbox")
@@ -92,9 +93,9 @@ export default {
   //   this.fetchPostData();
     // this.deletePost();
   // },
-  watch: {
-    deletePost(){}
-  },
+  // watch: {
+  //   deletePost(){}
+  // },
   data() {
     return {
       dropdownpost: false,
@@ -124,24 +125,33 @@ export default {
     openModal(){
       this.$refs.my_modal.visible = true;
     },
-    deletePost(pk, i){
-      // console.log('pkstpk::',pk);
-      // console.log('i', this.post_data[i]);
-      // let post_num = this.post_data[i];
-      // post_num.splice(0,1);
-      // this.post_data.post[i].splice(i, 1);
-      console.log('i',this.post_data);
-      // console.log('i',post_num);
+    deletePost(pk){
       let user_token = window.localStorage.getItem('token');
-      this.$http.delete('https://api.thekym.com/post/' + `${pk}`+ '/',
-       { headers: {'Authorization' : `Token ${user_token}`}})
-                .then(response=> {
-                  // post_num.splice(0,1);
-                  // console.log('i',this.post_data);
-                  // console.log('i',post_num);
-                  this.post_data.post[i].splice(i, 1);
-                })
-                .catch(error => console.log(error.response));
+      let user_nickname = window.localStorage.getItem('user_nickname');
+      let confirmPostDelete = confirm(`${user_nickname}` + '님, 정말 이 글을 삭제하시겠습니까?');
+      if ( confirmPostDelete === true ){
+        this.$http.delete('https://api.thekym.com/post/' + `${pk}`+ '/',
+          { headers: {'Authorization' : `Token ${user_token}`}})
+          .then(response=> {
+            let group_pk = window.localStorage.getItem('this_group');
+            this.$http.get('https://api.thekym.com/post/?group=' + `${group_pk}`,
+              { headers: {'Authorization' : `Token ${user_token}`} })
+              .then(response=> {
+                let data = response.data.results;
+                let post_data = []
+                data.forEach(item => {
+                  post_data.push(item);
+                });
+                this.data_list = post_data;
+              })
+          })
+          .catch(error => {
+            if (error.response.status === 403){
+              alert('작성자만 요청할 수 있는 작업입니다.');
+            }
+            console.log(error.response);
+          });
+      }
     },
     getMyGroupList(){
         let user_token = window.localStorage.getItem('token');
@@ -198,26 +208,6 @@ export default {
          this.openMywrite('prev');
       }
     },
-    deletePost(pk){
-      // this.$refs.delete_post_modal.visible = true;
-      // window.localStorage.getItem(pk);
-      console.log('pkstpk::',pk);
-      // console.log('i', this.post_data[i]);
-      // let post_num = this.post_data[i];
-      // post_num.splice(0,1);
-      // this.post_data.post[i].splice(i, 1);
-      // console.log('i',post_num);
-      let user_token = window.localStorage.getItem('token');
-      this.$http.delete('https://api.thekym.com/post/' + `${pk}`+ '/',
-       { headers: {'Authorization' : `Token ${user_token}`}})
-                .then(response=> {
-                  // post_num.splice(0,1);
-                  // console.log('i',this.post_data);
-                  // console.log('i',post_num);
-                  // this.post_data.post[i].splice(i, 1);
-                })
-                .catch(error => console.log(error.response));
-    },
     goGroup(pk){
       window.localStorage.setItem('this_group', pk);
       this.$router.push({ path: '/JointGroup/', query: { group: `${pk}` }});
@@ -251,14 +241,20 @@ body
 .card
   margin-bottom: 20px
 
-.card
-  margin-bottom: 20px
+.feed-box-wrapper
+  min-height: 80vh
 
 .dropdownhr
   margin: 5px
+.group-img-small-wrapper
+  width: 32px
+  height: 32px
+  overflow: hidden
+  border-radius: 10%
 
 .group-img-small
-  border-radius: 10%
+  width: 100%
+  min-height: 100%
 
 .group-small-list-group
   line-height: 38px
@@ -269,6 +265,17 @@ body
 
 .group-name
   color: $primary
+
+.img-user-64
+  background: #eee
+  width: 64px
+  height: 64px
+  overflow: hidden
+  border-radius: 50%
+
+.img-user-profile
+  width: 100%
+  height: 100%
 
 .btn-show-comment,
 .btn-show-like
